HIVE-24824: Create AcidMetricsService
