+++ ql/src/test/org/apache/hadoop/hive/ql/txn/compactor/CompactorTest.java
@@ -52,6 +52,7 @@ import org.apache.hadoop.hive.metastore.api.TxnAbortedException;
 import org.apache.hadoop.hive.metastore.api.TxnType;
 import org.apache.hadoop.hive.metastore.api.hive_metastoreConstants;
 import org.apache.hadoop.hive.metastore.conf.MetastoreConf;
+import org.apache.hadoop.hive.metastore.metrics.AcidMetricService;
 import org.apache.hadoop.hive.metastore.txn.CompactionInfo;
 import org.apache.hadoop.hive.metastore.txn.TxnCommonUtils;
 import org.apache.hadoop.hive.metastore.utils.TestTxnDbUtil;
@@ -134,6 +135,16 @@ public abstract class CompactorTest {
 
   protected void startCleaner() throws Exception {
     startThread(CompactorThreadType.CLEANER, true);
+  }
+
+  protected void runAcidMetricService() throws Exception {
+    TestTxnDbUtil.setConfValues(conf);
+    AcidMetricService t = new AcidMetricService();
+    t.setThreadId((int) t.getId());
+    t.setConf(conf);
+    stop.set(true);
+    t.init(stop);
+    t.run();
   }
 
   protected Table newTable(String dbName, String tableName, boolean partitioned) throws TException {
