+++ ql/src/test/org/apache/hadoop/hive/ql/txn/compactor/TestInitiator.java
@@ -865,7 +865,7 @@ public class TestInitiator extends CompactorTest {
   }
 
   @Test
-  public void testInitiatorMetricsEnabled() throws Exception {
+  public void testAcidMetricsEnabled() throws Exception {
     MetastoreConf.setBoolVar(conf, MetastoreConf.ConfVars.METRICS_ENABLED, true);
     Metrics.initialize(conf);
     int originalValue = Metrics.getOrCreateGauge(INITIATED_METRICS_KEY).intValue();
@@ -902,15 +902,15 @@ public class TestInitiator extends CompactorTest {
     List<ShowCompactResponseElement> compacts = rsp.getCompacts();
     Assert.assertEquals(10, compacts.size());
 
-    // The metrics will appear after the next Initiator run
-    startInitiator();
+    // The metrics will appear after the next AcidMetricsService run
+    runAcidMetricService();
 
     Assert.assertEquals(originalValue + 10,
         Metrics.getOrCreateGauge(MetricsConstants.COMPACTION_STATUS_PREFIX + TxnStore.INITIATED_RESPONSE).intValue());
   }
 
   @Test
-  public void testInitiatorMetricsDisabled() throws Exception {
+  public void testAcidMetricsDisabled() throws Exception {
     MetastoreConf.setBoolVar(conf, MetastoreConf.ConfVars.METRICS_ENABLED, false);
     Metrics.initialize(conf);
     int originalValue = Metrics.getOrCreateGauge(INITIATED_METRICS_KEY).intValue();
@@ -947,131 +947,13 @@ public class TestInitiator extends CompactorTest {
     List<ShowCompactResponseElement> compacts = rsp.getCompacts();
     Assert.assertEquals(10, compacts.size());
 
-    // The metrics will appear after the next Initiator run
-    startInitiator();
+    // The metrics will appear after the next AcidMetricsService run
+    runAcidMetricService();
 
     Assert.assertEquals(originalValue,
         Metrics.getOrCreateGauge(INITIATED_METRICS_KEY).intValue());
   }
 
-  @Test
-  public void testUpdateCompactionMetrics() {
-    Metrics.initialize(conf);
-    ShowCompactResponse scr = new ShowCompactResponse();
-    List<ShowCompactResponseElement> elements = new ArrayList<>();
-    elements.add(generateElement(1,"db", "tb", null, CompactionType.MAJOR, TxnStore.FAILED_RESPONSE));
-    // Check for overwrite
-    elements.add(generateElement(2,"db", "tb", null, CompactionType.MAJOR, TxnStore.INITIATED_RESPONSE));
-    elements.add(generateElement(3,"db", "tb2", null, CompactionType.MINOR, TxnStore.INITIATED_RESPONSE));
-    elements.add(generateElement(5,"db", "tb3", "p1", CompactionType.MINOR, TxnStore.DID_NOT_INITIATE_RESPONSE));
-    // Check for overwrite where the order is different
-    elements.add(generateElement(4,"db", "tb3", "p1", CompactionType.MINOR, TxnStore.FAILED_RESPONSE));
-
-    elements.add(generateElement(6,"db1", "tb", null, CompactionType.MINOR, TxnStore.FAILED_RESPONSE));
-    elements.add(generateElement(7,"db1", "tb2", null, CompactionType.MINOR, TxnStore.FAILED_RESPONSE));
-    elements.add(generateElement(8,"db1", "tb3", null, CompactionType.MINOR, TxnStore.FAILED_RESPONSE));
-
-    elements.add(generateElement(9,"db2", "tb", null, CompactionType.MINOR, TxnStore.SUCCEEDED_RESPONSE));
-    elements.add(generateElement(10,"db2", "tb2", null, CompactionType.MINOR, TxnStore.SUCCEEDED_RESPONSE));
-    elements.add(generateElement(11,"db2", "tb3", null, CompactionType.MINOR, TxnStore.SUCCEEDED_RESPONSE));
-    elements.add(generateElement(12,"db2", "tb4", null, CompactionType.MINOR, TxnStore.SUCCEEDED_RESPONSE));
-
-    elements.add(generateElement(13,"db3", "tb3", null, CompactionType.MINOR, TxnStore.WORKING_RESPONSE));
-    elements.add(generateElement(14,"db3", "tb4", null, CompactionType.MINOR, TxnStore.WORKING_RESPONSE));
-    elements.add(generateElement(15,"db3", "tb5", null, CompactionType.MINOR, TxnStore.WORKING_RESPONSE));
-    elements.add(generateElement(16,"db3", "tb6", null, CompactionType.MINOR, TxnStore.WORKING_RESPONSE));
-    elements.add(generateElement(17,"db3", "tb7", null, CompactionType.MINOR, TxnStore.WORKING_RESPONSE));
-
-    scr.setCompacts(elements);
-    Initiator.updateCompactionMetrics(scr);
-
-    Assert.assertEquals(1,
-        Metrics.getOrCreateGauge(MetricsConstants.COMPACTION_STATUS_PREFIX + TxnStore.DID_NOT_INITIATE_RESPONSE).intValue());
-    Assert.assertEquals(2,
-        Metrics.getOrCreateGauge(MetricsConstants.COMPACTION_STATUS_PREFIX + TxnStore.INITIATED_RESPONSE).intValue());
-    Assert.assertEquals(3,
-        Metrics.getOrCreateGauge(MetricsConstants.COMPACTION_STATUS_PREFIX + TxnStore.FAILED_RESPONSE).intValue());
-    Assert.assertEquals(4,
-        Metrics.getOrCreateGauge(MetricsConstants.COMPACTION_STATUS_PREFIX + TxnStore.SUCCEEDED_RESPONSE).intValue());
-    Assert.assertEquals(5,
-        Metrics.getOrCreateGauge(MetricsConstants.COMPACTION_STATUS_PREFIX + TxnStore.WORKING_RESPONSE).intValue());
-    Assert.assertEquals(0,
-        Metrics.getOrCreateGauge(MetricsConstants.COMPACTION_STATUS_PREFIX + TxnStore.CLEANING_RESPONSE).intValue());
-  }
-
-  @Test
-  public void testAgeMetricsNotSet() {
-    Metrics.initialize(conf);
-    ShowCompactResponse scr = new ShowCompactResponse();
-    List<ShowCompactResponseElement> elements = new ArrayList<>();
-    elements.add(generateElement(1, "db", "tb", null, CompactionType.MAJOR, TxnStore.FAILED_RESPONSE, 1L));
-    elements.add(generateElement(5, "db", "tb3", "p1", CompactionType.MINOR, TxnStore.DID_NOT_INITIATE_RESPONSE, 2L));
-    elements.add(generateElement(9, "db2", "tb", null, CompactionType.MINOR, TxnStore.SUCCEEDED_RESPONSE, 3L));
-    elements.add(generateElement(13, "db3", "tb3", null, CompactionType.MINOR, TxnStore.WORKING_RESPONSE, 4L));
-    elements.add(generateElement(14, "db3", "tb4", null, CompactionType.MINOR, TxnStore.CLEANING_RESPONSE, 5L));
-
-    scr.setCompacts(elements);
-    Initiator.updateCompactionMetrics(scr);
-    // Check that it is not set
-    Assert.assertEquals(0, Metrics.getOrCreateGauge(MetricsConstants.COMPACTION_OLDEST_ENQUEUE_AGE).intValue());
-  }
-
-  @Test
-  public void testAgeMetricsAge() {
-    Metrics.initialize(conf);
-    ShowCompactResponse scr = new ShowCompactResponse();
-    List<ShowCompactResponseElement> elements = new ArrayList<>();
-    long start = System.currentTimeMillis() - 1000L;
-    elements.add(generateElement(15,"db3", "tb5", null, CompactionType.MINOR, TxnStore.INITIATED_RESPONSE, start));
-
-    scr.setCompacts(elements);
-    Initiator.updateCompactionMetrics(scr);
-    long diff = (System.currentTimeMillis() - start)/1000;
-    // Check that we have at least 1s old compaction age, but not more than expected
-    Assert.assertTrue(Metrics.getOrCreateGauge(MetricsConstants.COMPACTION_OLDEST_ENQUEUE_AGE).intValue() <= diff);
-    Assert.assertTrue(Metrics.getOrCreateGauge(MetricsConstants.COMPACTION_OLDEST_ENQUEUE_AGE).intValue() >= 1);
-  }
-
-  @Test
-  public void testAgeMetricsOrder() {
-    Metrics.initialize(conf);
-    ShowCompactResponse scr = new ShowCompactResponse();
-    long start = System.currentTimeMillis();
-    List<ShowCompactResponseElement> elements = new ArrayList<>();
-    elements.add(generateElement(15,"db3", "tb5", null, CompactionType.MINOR, TxnStore.INITIATED_RESPONSE,
-        start - 1000L));
-    elements.add(generateElement(16,"db3", "tb6", null, CompactionType.MINOR, TxnStore.INITIATED_RESPONSE,
-        start - 100000L));
-
-    scr.setCompacts(elements);
-    Initiator.updateCompactionMetrics(scr);
-    // Check that the age is older than 10s
-    Assert.assertTrue(Metrics.getOrCreateGauge(MetricsConstants.COMPACTION_OLDEST_ENQUEUE_AGE).intValue() > 10);
-
-    // Check the reverse order
-    elements = new ArrayList<>();
-    elements.add(generateElement(16,"db3", "tb6", null, CompactionType.MINOR, TxnStore.INITIATED_RESPONSE,
-        start - 100000L));
-    elements.add(generateElement(15,"db3", "tb5", null, CompactionType.MINOR, TxnStore.INITIATED_RESPONSE,
-        start - 1000L));
-
-    // Check that the age is older than 10s
-    Assert.assertTrue(Metrics.getOrCreateGauge(MetricsConstants.COMPACTION_OLDEST_ENQUEUE_AGE).intValue() > 10);
-  }
-
-  private ShowCompactResponseElement generateElement(long id, String db, String table, String partition,
-      CompactionType type, String state) {
-    return generateElement(id, db, table, partition, type, state, System.currentTimeMillis());
-  }
-
-  private ShowCompactResponseElement generateElement(long id, String db, String table, String partition,
-      CompactionType type, String state, long enqueueTime) {
-    ShowCompactResponseElement element = new ShowCompactResponseElement(db, table, type, state);
-    element.setId(id);
-    element.setPartitionname(partition);
-    element.setEnqueueTime(enqueueTime);
-    return element;
-  }
 
   @Test
   public void compactTableWithMultipleBase() throws Exception {
